+++ functions/index.js
@@ -15,3 +15,4 @@ exports.stats = require('./stats').stats;
 exports.sealed = require('./sealed').sealed;
 exports.version = require('./version').version;
 exports.multiDeck = require('./multiDeck').multiDeck;
+exports.weirdo = require('./weirdo').weirdo;